fix 64-bit build warning

+++ src/priv/base/base/string.cpp
@@ -155,7 +155,7 @@ size_t GN::str2Floats( float * buffer, size_t count, const char * str, size_t st
     static pcrecpp::RE re( "\\s*([+-]?\\s*([0-9]+(\\.[0-9]*)?|[0-9]*\\.[0-9]+)([eE][+-]?[0-9]+)?)\\s*,?\\s*" );
 
     if( 0 == stringLength ) stringLength = strLen( str );
-    pcrecpp::StringPiece text( str, stringLength );
+    pcrecpp::StringPiece text( str, (int)stringLength );
 
     std::string substring;
     for( size_t i = 0; i < count; ++i )
